+++ engine/src/main/kotlin/com/kakao/actionbase/engine/catalog/CatalogLoader.kt
@@ -1,0 +1,7 @@
+package com.kakao.actionbase.engine.catalog
+
+import com.kakao.actionbase.engine.Engine
+
+interface CatalogLoader : AutoCloseable {
+    fun bind(engine: Engine)
+}
