--- engine/src/main/kotlin/com/kakao/actionbase/engine/runtime/CatalogLoader.kt
@@ -1,5 +1,0 @@
-package com.kakao.actionbase.engine.runtime
-
-interface CatalogLoader : AutoCloseable {
-    fun bind(engine: Engine)
-}
